Extract platform feedback subsystem into its own module
PlatformDashboardTabs.tsx had grown into a 2366-line god file holding
all 10 platform admin tabs plus the entire feedback triage subsystem
(submission detail wizard, item detail wizard, list views, and ~16
helpers).

Split the feedback subsystem (~1650 lines) into a dedicated
-components/feedback/ module:
- -feedback-helpers.tsx: label/format/build helpers + FeedbackMetaRow
- FeedbackTab.tsx: inbox + open-items list views
- FeedbackSubmissionDetailPanel.tsx: triage wizard
- FeedbackItemDetailPanel.tsx: item detail wizard

Shared table cell styles (TH/TD) moved to PlatformShell.tsx so both
the feedback module and the remaining simple tabs import one source.

PlatformDashboardTabs.tsx drops from 2366 to 659 lines and now only
holds the read-only table tabs. Pure code-move refactor: no behavior
change. tsc + eslint clean, 378 tests pass.
